Splitted professor/student routes into separate routes.
