+++ api/Routes/professorRouter.js
@@ -1,0 +1,6 @@
+import express from 'express';
+import { viewProfessorProfile } from '../Controllers/professorController.js';
+
+const professorRouter = express.Router();
+professorRouter.get('/professor/profile', viewProfessorProfile);
+export default professorRouter;
